refactor(targets): retire _common.py compat re-export module
Closes #232.

After PR #229 broke the original 1,557-LOC `_common.py` into focused
submodules (`_adapter`, `_combine`, `_driver`, `_intermediates`, `_io`,
`_shims`, `_writers`), `_common.py` survived as a 110-line back-compat
re-export shim. Time to delete it and force every consumer onto the new
homes — lingering shims attract new imports from people who don't read
the deprecation notice.

Changes:
- Delete `src/nhf_spatial_targets/targets/_common.py`.
- Redirect 97 test imports across `tests/test_targets_common.py`,
  `tests/test_cf_compliance.py`, and `tests/test_targets_swe.py` to
  the symbol's actual home (`_io`, `_shims`, `_combine`, `_writers`,
  or `_intermediates`).
- Update 4 caplog `logger="...targets._common"` refs (sca + swe tests)
  to `...targets._intermediates`, which is where the watched warnings
  (`predates`, `missing`, `mismatch`) actually originate now.
- Update stale doc comments in `io_nc.py`, `aggregate/era5_land.py`,
  `aggregate/_driver.py`, and `targets/_io.py` to point at the new
  homes.
- Update live architecture/reference docs
  (`nc-encoding-policy`, `python-patterns`, `transformation-pipeline`,
  `calibration-target-recipes`) for the same reason.

Historical design docs under `docs/superpowers/plans/` and
`docs/superpowers/specs/` keep their original `targets/_common.py`
references — they're snapshots of what was true at design time and are
excluded from the mkdocs build.

After this PR `grep -rn "targets\._common\|targets/_common" src/ tests/
docs/architecture/ docs/references/ docs/api/ docs/sources/` returns
zero hits.

Behavior unchanged. `tests/test_targets_common.py` (114 tests) +
`tests/test_targets_sca.py` + `tests/test_targets_swe.py` (68 tests)
all pass locally; full suite gated by CI.
